roberta-requirements/requirements.txt
torch>=1.7.0
pytorch-crf>=0.7.2
transformers>=4.5.0
numpy>=1.19.0
pandas>=1.1.0
tqdm>=4.50.0
matplotlib>=3.3.0
scikit-learn>=0.24.0
pytorch-lightning>=1.4.0
tensorboard>=2.4.0
jupyterlab>=3.0.0
notebook>=6.0.0
ipywidgets>=7.6.0
setuptools>=49.0.0
wheel>=0.35.0
gdown>=4.4.0
configparser>=5.0.0
sentencepiece>=0.1.95 